Rename docker container
