+++ patient-manager/app/enums.py
