+++ patient-manager/app/ressources/epi_pypulseq.seq
