+++ patient-manager/app/utilities/sequence_plot.py
